stable coin huge testcase

+++ contracts/eosdos/include/common/defines.hpp
@@ -21,7 +21,7 @@ using namespace eosio;
 using bytes   = std::vector<char>;
 using bytes32 = std::vector<char>;
 using address = name;
-using uint256 = uint64_t;
+using uint256 = uint128_t;
 using uint112 = uint64_t;
 using uint8   = uint8_t;
 using uint32  = uint32_t;
